@@ -680,7 +680,7 @@ func TestDuplicateTable(t *testing.T) {
 			t.Fatalf("%+v", ge)
 		}
 	} else {
-		t.Fatalf(("Threre should be error, when table name duplicates."))
+		t.Fatalf("Threre should be error, when table name duplicates.")
 	}
 }
 
@@ -737,7 +737,7 @@ func TestDuplicateTableWithSchema(t *testing.T) {
 			t.Fatalf("%+v", ge)
 		}
 	} else {
-		t.Fatalf(("Threre should be error, when table name duplicates."))
+		t.Fatalf("Threre should be error, when table name duplicates.")
 	}
 }
 
