@@ -2,8 +2,10 @@ package server
 
 import (
 	"bytes"
+	"cloud.google.com/go/bigquery/storage/apiv1/storagepb"
 	"context"
 	"fmt"
+	"google.golang.org/protobuf/types/known/anypb"
 	"io"
 	"strings"
 	"sync"
@@ -14,9 +16,8 @@ import (
 	"github.com/apache/arrow/go/v10/arrow/ipc"
 	"github.com/apache/arrow/go/v10/arrow/memory"
 	"github.com/goccy/go-json"
-	goavro "github.com/linkedin/goavro/v2"
+	"github.com/linkedin/goavro/v2"
 	bigqueryv2 "google.golang.org/api/bigquery/v2"
-	storagepb "google.golang.org/genproto/googleapis/cloud/bigquery/storage/v1"
 	"google.golang.org/genproto/googleapis/rpc/status"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -381,6 +382,10 @@ type writeStreamStatus struct {
 	tableMetadata *bigqueryv2.Table
 	rows          types.Data
 	finalized     bool
+	offset        int64
+	defaultStream bool
+	appendStream  storagepb.BigQueryWrite_AppendRowsServer
+	messageDesc   protoreflect.MessageDescriptor
 }
 
 func (s *storageWriteServer) CreateWriteStream(ctx context.Context, req *storagepb.CreateWriteStreamRequest) (*storagepb.WriteStream, error) {
@@ -392,10 +397,14 @@ func (s *storageWriteServer) CreateWriteStream(ctx context.Context, req *storage
 	if err != nil {
 		return nil, fmt.Errorf("failed to get table metadata: %w", err)
 	}
-	streamID := randomID()
-	streamName := fmt.Sprintf("%s/streams/%s", req.Parent, streamID)
+
+	writeStream := req.GetWriteStream()
+	if writeStream == nil {
+		return nil, fmt.Errorf("missing write stream")
+	}
+	streamName := writeStream.GetName()
 	createTime := timestamppb.New(time.Now())
-	streamType := req.GetWriteStream().GetType()
+	streamType := writeStream.GetType()
 	var commitTime *timestamppb.Timestamp
 	if streamType == storagepb.WriteStream_COMMITTED {
 		commitTime = createTime
@@ -418,6 +427,7 @@ func (s *storageWriteServer) CreateWriteStream(ctx context.Context, req *storage
 		datasetID:     datasetID,
 		tableID:       tableID,
 		tableMetadata: tableMetadata,
+		defaultStream: strings.HasSuffix(streamName, "_default"),
 	}
 	s.mu.Unlock()
 	return stream, nil
@@ -435,19 +445,46 @@ func (s *storageWriteServer) AppendRows(stream storagepb.BigQueryWrite_AppendRow
 	if err != nil {
 		return err
 	}
-	if err := s.appendRows(req, msgDesc, stream); err != nil {
-		return fmt.Errorf("failed to append rows: %w", err)
+
+	var streamStatus *writeStreamStatus
+	streamName := req.GetWriteStream()
+	s.mu.RLock()
+	streamStatus, _ = s.streamMap[streamName]
+	s.mu.RUnlock()
+	// create the default stream on demand
+	if streamStatus == nil && strings.HasSuffix(streamName, "_default") {
+		nameParts := strings.Split(streamName, "/")
+		if _, err := s.CreateWriteStream(context.Background(), &storagepb.CreateWriteStreamRequest{
+			Parent: strings.Join(nameParts[:len(nameParts)-2], "/"),
+			WriteStream: &storagepb.WriteStream{
+				Name: streamName,
+				Type: storagepb.WriteStream_COMMITTED,
+			},
+		}); err != nil {
+			return fmt.Errorf("failed to create default stream: %w", err)
+		}
 	}
+	s.mu.RLock()
+	streamStatus, _ = s.streamMap[streamName]
+	if streamStatus == nil {
+		s.mu.RUnlock()
+		return fmt.Errorf("stream %s not found", streamName)
+	}
+	s.mu.RUnlock()
+
+	streamStatus.appendStream = stream
+	streamStatus.messageDesc = msgDesc
 	for {
-		req, err := stream.Recv()
+		if err := s.appendRows(req, streamStatus); err != nil {
+			return fmt.Errorf("failed to append rows: %w", err)
+		}
+
+		req, err = stream.Recv()
 		if err == io.EOF {
 			break
 		}
 		if err != nil {
 			return err
-		}
-		if err := s.appendRows(req, msgDesc, stream); err != nil {
-			return fmt.Errorf("failed to append rows: %w", err)
 		}
 	}
 	return nil
@@ -468,56 +505,76 @@ func (s *storageWriteServer) getMessageDescriptor(req *storagepb.AppendRowsReque
 	return fd.Messages().ByName(protoreflect.Name(descProto.GetName())), nil
 }
 
-func (s *storageWriteServer) appendRows(req *storagepb.AppendRowsRequest, msgDesc protoreflect.MessageDescriptor, stream storagepb.BigQueryWrite_AppendRowsServer) error {
-	streamName := req.GetWriteStream()
-	s.mu.RLock()
-	var status *writeStreamStatus
-	if streamName == "" {
-		for _, s := range s.streamMap {
-			status = s
-			break
-		}
-	} else {
-		s, exists := s.streamMap[streamName]
-		if !exists {
-			return fmt.Errorf("failed to get stream from %s", streamName)
-		}
-		status = s
+func (s *storageWriteServer) appendRows(req *storagepb.AppendRowsRequest, streamStatus *writeStreamStatus) error {
+	if streamStatus.finalized {
+		return s.sendStorageErr(
+			streamStatus.appendStream,
+			streamStatus.stream.GetName(),
+			codes.InvalidArgument,
+			storagepb.StorageError_STREAM_FINALIZED,
+			"stream has been finalized and cannot be appended")
 	}
-	s.mu.RUnlock()
-	if status.finalized {
-		return fmt.Errorf("stream is already finalized")
+
+	var offset int64
+
+	if !streamStatus.defaultStream {
+		if req.GetOffset() != nil {
+			offset = req.GetOffset().GetValue()
+		}
+		if offset > streamStatus.offset {
+			return s.sendStorageErr(
+				streamStatus.appendStream,
+				streamStatus.stream.GetName(),
+				codes.OutOfRange,
+				storagepb.StorageError_OFFSET_OUT_OF_RANGE,
+				fmt.Sprintf("the offset is out of range, expected %d but received %d", streamStatus.offset, offset))
+		} else if offset < streamStatus.offset {
+			return s.sendStorageErr(
+				streamStatus.appendStream,
+				streamStatus.stream.GetName(),
+				codes.AlreadyExists,
+				storagepb.StorageError_OFFSET_ALREADY_EXISTS,
+				fmt.Sprintf("the offset is within the stream, expected %d but received %d", streamStatus.offset, offset))
+		}
 	}
-	offset := req.GetOffset().Value
+
 	rows := req.GetProtoRows().GetRows().GetSerializedRows()
-	data, err := s.decodeData(msgDesc, rows)
+	data, err := s.decodeData(streamStatus.messageDesc, rows)
 	if err != nil {
-		s.sendErrorMessage(stream, streamName, err)
+		s.sendErrorMessage(streamStatus.appendStream, streamStatus.stream.GetName(), err)
 		return err
 	}
-	if status.streamType == storagepb.WriteStream_COMMITTED {
+	if streamStatus.streamType == storagepb.WriteStream_COMMITTED {
 		ctx := context.Background()
 		ctx = logger.WithLogger(ctx, s.server.logger)
 
-		conn, err := s.server.connMgr.Connection(ctx, status.projectID, status.datasetID)
+		conn, err := s.server.connMgr.Connection(ctx, streamStatus.projectID, streamStatus.datasetID)
 		if err != nil {
-			s.sendErrorMessage(stream, streamName, err)
+			s.sendErrorMessage(streamStatus.appendStream, streamStatus.stream.GetName(), err)
 			return err
 		}
 		tx, err := conn.Begin(ctx)
 		if err != nil {
-			s.sendErrorMessage(stream, streamName, err)
+			s.sendErrorMessage(streamStatus.appendStream, streamStatus.stream.GetName(), err)
 			return err
 		}
 		defer tx.RollbackIfNotCommitted()
-		if err := s.insertTableData(ctx, tx, status, data); err != nil {
-			s.sendErrorMessage(stream, streamName, err)
+		if err := s.insertTableData(ctx, tx, streamStatus, data); err != nil {
+			s.sendErrorMessage(streamStatus.appendStream, streamStatus.stream.GetName(), err)
 			return err
 		}
+		if err := tx.Commit(); err != nil {
+			return err
+		}
+		if !streamStatus.defaultStream {
+			streamStatus.offset += int64(len(rows))
+		} else {
+			offset = -1
+		}
 	} else {
-		status.rows = append(status.rows, data...)
+		streamStatus.rows = append(streamStatus.rows, data...)
 	}
-	return s.sendResult(stream, streamName, offset+int64(len(rows)))
+	return s.sendResult(streamStatus.appendStream, streamStatus.stream.GetName(), offset)
 }
 
 func (s *storageWriteServer) sendResult(stream storagepb.BigQueryWrite_AppendRowsServer, streamName string, offset int64) error {
@@ -538,6 +595,37 @@ func (s *storageWriteServer) sendErrorMessage(stream storagepb.BigQueryWrite_App
 			Error: &status.Status{
 				Code:    int32(codes.Internal),
 				Message: err.Error(),
+			},
+		},
+	})
+}
+
+func (s *storageWriteServer) sendStorageErr(
+	stream storagepb.BigQueryWrite_AppendRowsServer,
+	streamName string,
+	errCode codes.Code,
+	storageErrorCode storagepb.StorageError_StorageErrorCode,
+	errMessage string) error {
+	storageErrBy, err := proto.Marshal(&storagepb.StorageError{
+		Code:         storageErrorCode,
+		Entity:       streamName,
+		ErrorMessage: errMessage,
+	})
+	if err != nil {
+		return err
+	}
+	return stream.Send(&storagepb.AppendRowsResponse{
+		WriteStream: streamName,
+		Response: &storagepb.AppendRowsResponse_Error{
+			Error: &status.Status{
+				Code:    int32(errCode),
+				Message: errMessage,
+				Details: []*anypb.Any{
+					{
+						TypeUrl: "type.googleapis.com/google.cloud.bigquery.storage.v1.StorageError",
+						Value:   storageErrBy,
+					},
+				},
 			},
 		},
 	})
@@ -679,6 +767,7 @@ func (s *storageWriteServer) GetWriteStream(ctx context.Context, req *storagepb.
 }
 
 func (s *storageWriteServer) FinalizeWriteStream(ctx context.Context, req *storagepb.FinalizeWriteStreamRequest) (*storagepb.FinalizeWriteStreamResponse, error) {
+	// TODO(dm): return an error when trying to finalize a default stream
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 	status, exists := s.streamMap[req.GetName()]
